fix(core:runtime): deadlock
I've added extra info to runtime's `startStream` method.
It was port points (which I now call port addresses).
With that data in that method I could log the flow, which I did.
As it turns out channels for summator array-inport were generated
in different order (because source of that data is map - of keys
not deterministic).

When runtime was trying to write message to summator second array slot,
that operator was blocked at first slot. Solution is to make all logic
completely async - order should not be important. Simply adding goroutine
on top of blocking operation was a solvation.

Last part I can't understand (and that's scares me) - to me it looks like
a couple commits ago I had almost same setup (but no extra data and logging
in runtime) but it wasn't working...