Reproduce issues with struct builders
In the feature-std-image branch such struct builder code returns an
error like this.

It seems to be caused by the use of nested type parameters like
Struct<stream<repro.Foo>> as simply Struct<repro.Foo> is not enough to
trigger it.

This is a minimal repro. It could have something to do with issue #566
or be subtly different. We might need the same treatment for generic
frames.